pyproject.toml
[project]
name = "ai-deep-research-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "alexnb", email = "anestorbergmann@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "langchain-core>=0.3.74",
    "langgraph>=0.6.4",
    "openai==1.93.0",
    "pydantic-ai>=0.6.2",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ai_deep_research_assistant"]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.8",
]
